Provide basic support for running over tramp
Mostly the changes were about replacing shell-commands with replacement
commands that respect TRAMP connections (such as
process-file-shell-command).

In addition a few places were assuming that justfile path would be
local. We make sure to use tramp to translate any remote path to local.

Final change is the addition of --color=never switch to make sure we
disable colors unconditionally (in my case this was broken due to how
justfile was structured).